refactor(TASK-0006): ajuste no código
