+++ src/styles/@types/styled.d.ts
@@ -1,9 +1,9 @@
-/* eslint-disable @typescript-eslint/no-empty-interface */
 import 'styled-components'
-import { defaultTheme } from '../themes/default'
+import { defaultTheme } from '../theme/default'
 
 type ThemeType = typeof defaultTheme
 
 declare module 'styled-components' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
   export interface DefaultTheme extends ThemeType {}
 }
