+++ src/styles/theme/default.ts
@@ -3,7 +3,7 @@ export const defaultTheme = {
   black: '#000',
   gray: '#495057',
 
-  'green-300': '#3DCCB7;',
+  'green-300': '#3DCCB7',
   'green-500': '#008F79',
   'green-700': '#004238',
 } as const
